Try restoring the old values of x1 and x2
I had changed these to zero, because I think they should be zero
(because the benchmark accumulates products into them using `+=`), but
I want to minimize test diffs so I'll use the old data for now. I had
also omitted the "in" side of the copy for x1 & x2, because they truly
are just outputs and do not need to be inputs too, but again I want to
match the original behavior.