Optimize GraphColoring::insert_conflicts
A perf analysis revealed that `minimize-regs` is the slowest
optimization when trying to compile large systolic arrays (~20-30 second
for 8x8).

The culprit was `insert_conflicts` which was taking roughly ~90% of the
of the pass's runtime.

Implemented two optimizations:
1. itertools::tuple_combinations to enumerate fewer edges.
2. Use petgraph::matrix_graph which optimizes insertion and edge storage

Systolic arrays are probably the worst case for minimize-regs since
everything has an edge to edge (to a first approximation).