Created TimeLockedValue and execute_mut
- TimeLockedValues are only set up by Primitives, their
execution/handling/coordination is the interpreter's job. This
will be handy for pipelining stuff (that we are not yet up to)

- Decentralized the Execute trait. Added default implementations for
execute to both ExecuteBinary and ExecuteUnary, and now also for
ExecuteStateful. No longer defininging Execute trait for ExecuteBinary.
So it seems like the interpreter will have to do some level of pattern
matching...