Wrote reg_seq test case
- Also modified if_reg to follow the same convention/goal: if reg0 and
reg1 are executed sequentially, then the done of reg0 should be 0 in the
output. That is because group B runs in the cycle after group A. If
group A's done is dependent on reg0's done, then group B runs a cycle
after reg0's done is high, meaning reg0's done must then be low.

- Currently the interpreter does not have support for this done
coordination, but i think it makes sense to put this in the tests!