Two big updates:
- Had an error in add_feeding.futil of width mismatch. The reason?
Values::new(bitwidth) used BitVec.with_capacity(), which has no bearing
on length; but check_widths() checks lengths, so unitialized Values
created with new() of a certain bitwidth cause errors. Fixed this.

- New error: from add_feeding.futil. Compiles and has output of
vcd_json, but is stuck in an infinite loop with the interpreter. Seems
like the [done] signals of the 2 registers might be out of step?