Remove fud's accidental dependency on calyx-py (#597)
At some point, fud started using a float-to-fixed conversion utility
from the calyx-py library. I think this was a mistake because calyx-py
is really just for Calyx code generators (frontends); it's not, like, a
general-purpose library for all things Calyx. And because this
dependency was undeclared, it meant that someone following our setup
instructions would end up with a "broken" fud and would need to go back
and install calyx-py as well.

This duplicates that utility function, which removes the accidental
dependency. I think it's not too much of a sin to copy & paste the code
because it's so little code. I also thought the original docstring was
somewhat inaccurate, so I tried a slightly better one.