Update attribute documentation (#599)
* Document shorthand attributes from #566

Update the documentation for attributes to mention the new `@foo`
syntax, which, as of #566, is a shorthand for the common case of `@foo(1)`.

* Document `@reset` attribute

This signal is available as of #579, and it's kind of a sibling to `@go`
and `@done`. It's now described alongside those. I also added some extra
text about what `@go` and `@done` really mean to programmers, although
this obviously bears further explanation.

* Document `@clk` attribute

For completeness on this list of all the attributes we use.