Trying to sort out register behavior
Reg's [out] and [done] ports are asserted exactly 1 cycle after
[write_en] is asserted. But does that mean a register needs 1 do_tick()
to express its output ports, or 2? From interpret_group.rs it seems like
a register is executed and then, in the same interpret_assignment(),
working_env.do_tick() is called, and so then the TLV is unwrapped.
Looking at TLV do_tick(), the TLV's count is decremented, and then
there's an attempt to unwrap it. Well since this do_tick() is called in
the same method call (interpret_assignment()) as the TLV of count 1 is
produced, that should mean that the register's [out] and [done] ports
are ready at the end of the first cycle/start of the second.
Confusing... is that correct behavior? And does that mean we need just 1
do_tick() call, or 2?