Do tick prims (#609)
Added some cycle-sentience to primitives




* Trying to sort out register behavior

Reg's [out] and [done] ports are asserted exactly 1 cycle after
[write_en] is asserted. But does that mean a register needs 1 do_tick()
to express its output ports, or 2? From interpret_group.rs it seems like
a register is executed and then, in the same interpret_assignment(),
working_env.do_tick() is called, and so then the TLV is unwrapped.
Looking at TLV do_tick(), the TLV's count is decremented, and then
there's an attempt to unwrap it. Well since this do_tick() is called in
the same method call (interpret_assignment()) as the TLV of count 1 is
produced, that should mean that the register's [out] and [done] ports
are ready at the end of the first cycle/start of the second.
Confusing... is that correct behavior? And does that mean we need just 1
do_tick() call, or 2?

* Changed primitives API

- For comb, nothing is diff in interpret_group
- For stateful, we only read from do_tick(). Execute() will save new
inputs, but do_tick() will decide (based on internal cycle count of
primitive) when to return the stored values of a stateful component
- No test will pass here yet. First step: write do_tick() of reg

* Starting reg interface

- Look at reg.futil in tests/primitives; that's an infinite loop
- Next step: unit testing reg interface, to see if prob is interface or
interpret_group

* reg test working in cargo test, so must be interpreter that is giving infinite loop

* Reg interface seems to work

- reg.futil test works (do it with fud)

- problem was/ returning a done signal even when cycle_count was 0, i
think. not sure why

* mem test, just d1, do that then do d2, d3, etcc

* Last_idx and update's idx not in sync

- For memd1, see the mem2.futil test (run through fud). Seems like group
[write2] and [dont_write] are interpretted in the same loop? Though they
are seperate groups. So the update with 3 as addr0 is taken after
execute(addr0=2...) is run, so the assert in [do_tick] panics

* Is [write_en] the [go] for mem?

- For [mem2.futil], reg0 gets 0 written to it. but that's from addr0 = 2
in [dont_write], the group before [write_en_low_outputs]. why is that?

* added some debug printing, seems like mem0 is ticked 3 times per group though each group should take 1 cycle. Anyways bug still persists, update holds addr0 = 3 while last_idx = 2, from dont_write

* Fixed the memory bug

But really what it was was I didn't understand [read_data] is supposed
to be combinational w.r.t [addr0]. The way to accomplish this is to have
both [execute] and [do_tick] return. So [execute] always returns based
on old value at [addr0], whereas [do_tick] only returns (w/ a high done
val) if [write_en] is high

* added primtest for std_mem_d1 and more detailed mismatch width panic

* wrote tests in primitives.rs for d2, d3, d4

* do_tick for all mem, should figure out what's wrong w runt

* did mult

* have dive

* removed done port from tests

* some documentation

* got rid of print in interp

* ready to PR

* clippy

* made mult truly 3 cycles

* changed cycle_count to write_en, and have [flickering_go] test to see incorrect interpret_group behavior. But weird guard behavior on group's done signal

* mult_pipe still succumbs to flickering go despite modification to execute; see flickering_go test

* fixed cargo tests for mult + div so they align to flickering-go behavior, and for runt test just needed to cargo build

* added correct flickering_go expect

* clippy

* comments before done port comment

* Added low done to stateful components run with low write_en

- The infinite loop bug was a problem with [get_cells] in
[interpret_group]; the problem was it didn't filter out duplicates

Co-authored-by: Alma Thaler <at564@cornell.edu>