Correctly report stdout on command failure
In debug mode (-vv), we don't have stderr captured (it is streamed
directly to the terminal). However, we were *also* giving up on
reporting the stdout, which we do still capture. It would be nicer to
also stream this in real time, but at least now it's possible to see the
messages from stdout.