[fud] Make SSH optional in Xilinx synthesis (xclbin) stage (#850)
* First attempt to refactor Xilinx SSH stuff

To use the existing `RemoteExecution` utility class instead of bespoke
SSH interaction. Hopefully, this will make it easier to let the stage
optionally work locally instead of via SSH.

* Decouple RemoteExecution from device_files

Eliminating the simpler one-file-sending API. Still need to test the
Xilinx stage to make sure it still works, but this avoids looking for
device_files in the xclbin stage.

* Further decouple from target_name

* Fix reference to mktmp

* No explicit input_files argument

Is this a sin? I don't *think* it is a problem to use closure state to
set up a stage like this, because the parameter is invariant here.

* One more closure parameter

* Slightly simplify path manipulation

* One step per copied file

* Attempt type-correctness

* One more type fix

* First attempt at local execution

* Fix new utility invocation

* Total wild attempt at local file reading

* Directory conversions

* Proper stage to read file

* Add a description

* Proper file-copying stages

* Little fixes for copy_files

* Correctly report stdout on command failure

In debug mode (-vv), we don't have stderr captured (it is streamed
directly to the terminal). However, we were *also* giving up on
reporting the stdout, which we do still capture. It would be nicer to
also stream this in real time, but at least now it's possible to see the
messages from stdout.

* Attempt to avoid need for tempdir

* Simplify sandboxing

Now `_sandbox` is more parallel to `open_and_send` on the remote path.

* Fix source wrapping

* Refactor `shell` utility

Hopefully this is a little more legible.

* Option to not capture stdout

Slightly better for debugging: verbose mode can stream both stdout *and*
stderr to the terminal!

* Avoid a None error

* One more None fix

* Tolerate plain values instead of Sources

* Source-wrap the sandbox calls

* Avoid double-wrapping

* Remove some suspicious cleanup

I don't completely understand what was going on with these lines, but
they had some problems:

- They would attempt a "self-move" and crash by moving a file into the
  directory it already existed inside.
- They use plain string manipulation for shell execution, which is
  generally dangerous and makes me nervous (especially with `rm`
  involved).

* Accept Black formatting

I don't love all of its choices, nor the 88-column option, but I don't
have to.

* Simplify old setup instructions

* Document running of synthesis stuff

* Remove unused import

* Start documenting real execution

* Document xclbin generation

* Document xclbin config options

* Docstrings

* Clarify use of `remote` in stats path