[fud] Make SSH optional in Xilinx emulation stage (#852)
* Start refactoring generic sandbox utility

LocalSandbox is a new alternative to RemoteExecution for running
commands locally. It's not a drop-in replacement (maybe that should come
in the future), but it offers morally equivalent functionality.

* Fix some configuration stuff

* Improve an error message

I don't know why "provide an input file" was suggested, so I took that
out. Also, properly interpolate the variable into the example flag.

* Document what we know so far

* Black formatting

* Simplify emulation command execution

And also make them work locally, with a utility.

* Avoid some hard-coding

* Document wdb stage configuration

* Fix for name change from #851