[fud] Start some very tiny tests for Xilinx toolchain (#866)
* First attempt at Xilinx compilation test

* Attempt actual execution test

* Many, many fixes

* Separate output directories

* Add input data for Xilinx test

* First stab at expected output

`compile` output is currently empty; that's not great. `emulate` output
has a bunch of junk and should probably not contain the running time.
But here we are for now.

* Add emconfig.json

* Suppress output from setup scripts

Turns out these were responsible for much of the output junk.

* Attempt to avoid emulation logs

Something is wrong here. For no reason I can see, I started running
into:

    [XRT] ERROR: kernel '_source' not found

And I had to fix this with the janky Python assignment you see below. At
the same time, the `output.log` stopped showing up. Also, the docs say
that the emulation logs *always* go to `emulation_debug.log`, but I have
never seen that file created.

* Run Black

I don't like this space wasted on just brackets...

* Optionally ignore log file

The problem with PyOpenCl appears to be here:
https://github.com/inducer/pyopencl/blob/a56281afc563e6be448cd8b2b2b69bc3afc59d78/pyopencl/__init__.py#L465

I still don't know why this worked at all. And I don't know why the log
file just vanishes sometime; maybe just because there's nothing to log.

* Update outputs for log-suppression work

* Run Black again

* Grep out some more log lines

This addresses these logged lines that I could not seem to disable through any
Xilinx configuration means (there are so many, and it's possible I
haven't found the right incantation):

    INFO::[ Vitis-EM 22 ] [Time elapsed: 0 minute(s) 17 seconds, Emulation time: 0.109057 ms]
    Data transfer between kernel(s) and global memory(s)
    Toplevel_1:m0_axi-HBM[0]          RD = 0.125 KB               WR = 0.000 KB

* Fix some quoting business

* Move emulation time to log

Instead of putting in the proper output. It kinda seems like the log is
where this belongs anyway, and this makes the output deterministic.

* Update outputs

Co-authored-by: Rachit Nigam <rachit.nigam12@gmail.com>