Stage initialization cannot use config.
Instead _define_steps gets access to the `config` object. This change,
in addition to the delayed calling of Stage.setup() means that stages
can require certain options to be defined on the configuration for the
stage to be correctly executed.

Previously, stages could not do this, instead having to specify default
values for all options, because stage construction would use the config
object.