Fix the Xilinx tests to work with recent fud (#947)
* Fix the Xilinx xclbin generation test

In light of the problems in #867 and #946, the only way to make .xclbin
generation work in fud is to combine *both* `-o foo.xclbin` *and*
`--to xclbin`. This makes the relevant test do exactly that.

(As a quick reminder, the reason we can't do `--to xclbin` by itself is
that fud doesn't yet support binary output on stdout. So we must produce
a file instead.)

* Add `--to xclbin` to the Xilinx docs