Some miscellaneous builder improvements (#1296)
* [builder] Add width inference for seq_mem_d1

* [builder] Use sets for `par` control

...and also implicitly import the newish `memories` header, which has
the seq_mem components.

* Fix an error message

This thought it was a format string but it was not.

* [builder] Support building `comb group`s

* [builder] Simple builder for `while` control

* [builder] Catch malformed RHS in assignment

I wrote in my builder `mem.addr0 = reg` instead of `mem.addr0 =
reg.out`. The confusing error happened on code *generation*. This makes
the problem a bit more visible.

* [builder] Support `if` control statements

* Width inference for std_wire

* Add equality expressions and invoke statements

* Implicitly instantiate component builders

...and also support != for not-equals.

* Add a `program.import_` utility

* Only import `core` by default

We now *implicitly* import the `binary_operators` whenever you use
`add`. Of course we should probably add more of these implicit imports
for other stuff... and someday rethink the whole interface for importing
stuff from libraries.

* Fix references to the group within `GroupBuilder`

* Import `binary_operators` in NTT generator

* *Actually* only add imports once