Disable inlining in Verilator
I know this is not a great idea, and maybe we can fix it. But for now,
this works around a problem where Verilator doesn't like the relative
reference to the memory array in a `$readmemh` because the `main`
component has been inlined. The message looks like this:

%Error: Internal Error: /var/folders/9l/69gbl1wn5f3fpc47l_b4fslm0119xg/T/tmpo2r20t1i:488:39: ../V3LinkDot.cpp:2700: Couldn't resolve inlined scope 'main' in: TOP.main
  488 |     $readmemh({DATA, "/mem.dat"}, mem.mem);

We could consider decorating the main module with a /*verilator
no_inline_module*/ comment, but that seems annoying too. This will have
to do for now.