Proposal: Let Verilator generate its own `main` function (#1335)
* Fix a missing Python format string

This error message was less helpful before.

* Extremely crude attempt to rip out our testbench

This doesn't work yet, because the default Verilator `main` testbench
doesn't know how to start/stop Calyx components (i.e., it doesn't know
about go/done).

* Reuse the Icarus testbench for Verilator

We should of course someday move this out of the `icarus` subdirectory.

* Fix some bugs in the Verilog testbench

These are places where Verilator was more strict about things than
Icarus, apparently. The "width" specifier for `cycle_count` was
backward, and we need to use `longint` to get a 64-bit cycle count limit
(`int` is apparently only 32 bits, and Verilator doesn't like the
mismatch in the comparison).

* Disable inlining in Verilator

I know this is not a great idea, and maybe we can fix it. But for now,
this works around a problem where Verilator doesn't like the relative
reference to the memory array in a `$readmemh` because the `main`
component has been inlined. The message looks like this:

%Error: Internal Error: /var/folders/9l/69gbl1wn5f3fpc47l_b4fslm0119xg/T/tmpo2r20t1i:488:39: ../V3LinkDot.cpp:2700: Couldn't resolve inlined scope 'main' in: TOP.main
  488 |     $readmemh({DATA, "/mem.dat"}, mem.mem);

We could consider decorating the main module with a /*verilator
no_inline_module*/ comment, but that seems annoying too. This will have
to do for now.

* Fix executable name

Apparently, this is "V" plus the name of the top module. Now our top
module is called "TOP".

* Fix cycle count parsing

Our new `$display` method of outputting the cycle count leaves a bunch
of white space. So we can use the same parsing logic as the Icarus
stage.

* Downgrade unknown option error to warning

Co-authored-by: Rachit Nigam <rachit.nigam12@gmail.com>