Fix Xilinx `kernel.xml` generation after a dependency broke it (#1529)
* Clean up Xilinx tests

I'm not sure how, but it seems like these symlinks got rearranged into the
wrong directories in #1153. The `compile/` and `emulate/` directories are just
supposed to contain outputs, and the inputs are supposed to be in the main
test directory.

* Fix XML generation

Adapt to quick_xml >= 0.28.0.

* Fix Xilinx Verilog compilation tests

Put the expect files into their own proper directory, and point the Runt
config at that. The `compile/` directory is supposed to be empty files,
just marking the success of the Vivado execution.

* Add empty placeholders for compile tests

* Add tests for kernel.xml generation