Disable verification when compiling for Xilinx
We need `--disable-verify` for Xilinx-compatible Verilog. I
unfortunately learned this the very hard way by rediscovering it: all my
Xilinx simulations would hang indefinitely, and when looking at the log,
it was always because of a "multiple assignments" crash on something
seemingly innocuous. I got as far as breaking out gtkwave and looking at
the individual guard variables to see where the conflict came from.

There was no conflict. I think this gets triggered only because the very
first few moments of simulation make all the registers "x"s, before they
are reset for the first time. That's when the conflict triggers. We
don't care about this, so the thing to do is to disable the checks.

Of course, I learned belatedly that we already tried to do this for the
`synth-verilog` stage. But the Xilinx-specific version embedded in the
xclbin stage did not get the same treatment. :( Fixed now!