Rework to greatly simplify Xilinx execution (#1530)
* Start a new xclbin utility

It's a replacement for `fud_pynq_script` that is meant to be invoked as
a subprocess, not run directly from Python. This makes it more
replaceable and also allows us to source the necessary scripts before
invoking it.

* Refactor

* Actually use xclrun in Xilinx stage

* Avoid the need for directory-change shenanigans

* Put xclrun output in a file

* Generate emconfig.json for execution

* Document the new execution workflow

* Document the xclrun utility

* Runt tests use the new FPGA way

* Break a long line

* Remove vestigial `fud_pynq_script.py`

* Disable verification when compiling for Xilinx

We need `--disable-verify` for Xilinx-compatible Verilog. I
unfortunately learned this the very hard way by rediscovering it: all my
Xilinx simulations would hang indefinitely, and when looking at the log,
it was always because of a "multiple assignments" crash on something
seemingly innocuous. I got as far as breaking out gtkwave and looking at
the individual guard variables to see where the conflict came from.

There was no conflict. I think this gets triggered only because the very
first few moments of simulation make all the registers "x"s, before they
are reset for the first time. That's when the conflict triggers. We
don't care about this, so the thing to do is to disable the checks.

Of course, I learned belatedly that we already tried to do this for the
`synth-verilog` stage. But the Xilinx-specific version embedded in the
xclbin stage did not get the same treatment. :( Fixed now!

* Update expected emulation outputs

I think these are now actually... correct? 😱

* Shorter path to xclrun utility

* Document minimal xclrun invocation