Try setting up Depot for Docker builds (#1538)
* Try setting up Depot for Docker builds

Depot is awesome because it supports multi-platform images natively.
This is an attempt to invoke Depot on every push to our main branch,
updating the image for the `:latest` tag. (Release-based tags will have
to come later...)

* Add Depot badge to README

The Depot folks have generously donated a sponsored account for us, and
they asked us to acknowledge them with a README badge.

* No need to checkout

Apparently, the build action uses "git context" by default, meaning that
BuildKit is responsible for checking out the right revision.