Prefer ADD over `RUN git clone` in Dockerfile
We now use Docker's built-in support for fetching git repositories.
The functional change is that we no longer clone the Calyx git repo
inside the container; instead, we use the source code from the context
of the build. This means if you clone the repo and type `docker build`
yourself, you get an image corresponding to the files you have on your
machine, rather than a specific tag. This lets us build the `:latest`
tag on every push.