Fixes for the new Depot action (#1539)
* Prefer ADD over `RUN git clone` in Dockerfile

We now use Docker's built-in support for fetching git repositories.
The functional change is that we no longer clone the Calyx git repo
inside the container; instead, we use the source code from the context
of the build. This means if you clone the repo and type `docker build`
yourself, you get an image corresponding to the files you have on your
machine, rather than a specific tag. This lets us build the `:latest`
tag on every push.

* Fix tag for Depot push

Oops; the "bare" tag like the old one was trying to push to the central
Docker Hub instead of GitHub's container registry.