Small tweaks to fud2 (#1879)
* More consistent Ninja variable names

I was inconsistent about whether to use snake_case or kebab-case; this
standardizes us on the latter (and brings a few other names in line with
each other).

* Rename resources directory

The old name, `data`, was ridiculously vague and I don't know why I ever
thought it was a good idea. `rsrc` is slightly better.

This also changes the config option name; if you were using fud2
already, you will need to change this.