initial support for multiple errors
introduce `DiagnosticContext` that passes can optionally add errors to.
a `DiagnosticPass` is a pass that can produce an iterator of errors.
when a pass is a `DiagnosticPass`, you can register with
`.register_diagnostic()`. this returns the first error that the pass
accumulates, if any.

also update calyx-lsp to use new diagnostic passes so that it can
report multiple errors in a file.