[fud2] Even more confidently delete temp dir
In #2058, I made fud2 clean up the `.fud2` temporary directory even when
Ninja exited with an erorr. Unfortunately, that doesn't cover other
possible errors, such as when launching the Ninja process fails
altogether:
https://github.com/calyxir/calyx/pull/2058#issuecomment-2136078085

Rather than patching one error at a time, this uses `Drop` to be sure we
do this when any error occurs. It also means it will get deleted even
when we panic, which is *definitely* overkill, but I guess it's not a
big deal.