Round Robin oracle and eDSL (#2177)
This PR makes progress towards #1810. It implements the python oracle
for PIFOs generalized to n flows, now known as Round Robin queues. Just
as with the PIFO, if a flow falls silent, the remaining flows will take
their turns. That flow effectively skips its turn. To re-generate the
test files with 20000 commands and a max length of 16, type in the
command line after navigating to the directory calyx/calyx-py/calyx

```
./gen_queue_data_expect.sh
```

Additionally, this PR also implements the Calyx version of Round Robin
queues in rr_queue.py. This was originally supposed to be its own PR,
but I thought it might be more complicated if I branched off a branch.
To run these tests, type in the command line
```
runt -i "rr_queue"
```

---------

Co-authored-by: Cassandra Nicole Sziklai <cns58@havarti.cs.cornell.edu>
Co-authored-by: Anshuman Mohan <anshumanmohan@live.com>
Co-authored-by: Anshuman Mohan <10830208+anshumanmohan@users.noreply.github.com>