[fud2] update path finding to support multi state input and output (#2134)
This PR begins progress towards addressing #1958. It updates the
algorithm used by `fud2` for finding paths from input files to output
files to take in operations which have many inputs and outputs.

The behavior concurs with the current `fud2` in some but not all cases
and does not maintain `fud2`'s always finding the shortest path of
operations.

Specifying operations themselves and emitting the Ninja file still
assume single state inputs and outputs. This will be changed in further
PRs.