[fud2] Fix planner so that it doesn't rely on the order operations are defined (#2211)
There was a small bug with the planner where it would stop searching
when a through op appeared anywhere in the `ops` map. It really should
only stop if an op is one edge away from the current state. This worked
before because I guess the `ops` map happened to be ordered in such a
way where things worked.

Fixes #2209.