Cider DAP State Inspection + Metadata Update (#2238)
This branch provides groundwork for state inspection while debugging
Calyx programs in VSCode. It currently displays all the cells in the
program and provides methods used to obtain port information. Port
information is not currently displayed due to a still unclear issue.

This branch also updates the metadata table so that each entry takes on
the following form:
component_name.group_name : path/to/file start-end
this update fixes an issue where the metadata pass was not recognizing
Calyx programs where two unique components each have a group with the
same name and displays the range the group is on rather than just the
start line.

---------

Co-authored-by: Serena <smd334@cornell.edu>