Add a `.yxi` to `kernel.xml` generator (#2229)
In theory, this should be a step in allowing us to deprecate the Xilinx
backend. In particular this should replace
[`xml.rs`](https://github.com/calyxir/calyx/blob/main/calyx-backend/src/xilinx/xml.rs).
Not yet ready to totally be deprecated as end-to-end parity isn't there
yet.

There were a few minor outstanding questions that stem from a lack of
detail in the [XRT `kernel.xml`
spec](https://docs.amd.com/r/en-US/ug1393-vitis-application-acceleration/RTL-Kernel-XML-File).
But I don't think any questions are blocking.

One difference from the existing `xml.rs` is that this generator does
not assume a `timeout` argument to our function call, as the AXI
generators don't currently support such a timeout argument anyways.