Update `.expect` interface (v2) (#2231)
This PR makes progress towards
https://github.com/calyxir/calyx/issues/2170. It applies the new expect
interface to all queues. These queues call queue_call, which has the
updated code. Additionally, code for the oracles had to be updated, in
particular the method operate_queue(). The value 2^32 - 2 is given to
pushes, and the value 2^32 - 1 for errors generated from underflow or
overflow.

This is the new version of #2227, because of branch environment issues.

---------

Co-authored-by: Anshuman Mohan <anshumanmohan@live.com>