Massage away 2-way PIFOs (#2234)
This PR makes progress towards #2226 by updating variable names and
comments to account for the fact that round robin PIFOs can make n-way
trees. Additionally, it implements a PIFO tree and a more complicated
tree (`complicated_tree.py`) using round robin and strict queues. This
accounts for the first three checkboxes of #2226. The docs have been
updated to reflect the new PIFO style in #2223. To run the relevant
tests,
```
runt -i "queues"
```

(For sdn.py, I removed the stats component stuff because I thought it
was mentioned that this wasn't important and so the new style PIFOs are
not equipped to do stats. If it needs to be there, let me know.)

---------

Co-authored-by: Anshuman Mohan <anshumanmohan@live.com>