Update documentation: New style PIFOs, RR, and Strict (#2237)
This PR makes progress towards
https://github.com/calyxir/calyx/issues/2191 by adding documentation
about round robin and strict queues. Additionally, it completes the last
checkbox of https://github.com/calyxir/calyx/issues/2226 by updating the
documentation to use the new style of PIFOs. Recently, it also updates
the documentation to reflect the removal of peek (#2241). Because of a
git mistake carried forward, this PR supersedes #2223.

---------

Co-authored-by: Anshuman Mohan <anshumanmohan@live.com>