[fud2] improve error message for parser errors (#2258)
It does what it says on the tin. The error messages for the parser won't
be good, but they are better than rust panics and I think become
functional enough.

One interesting thing to note is I chose to print full paths when
referencing files. This is because the directory of scripts is often not
where `fud2` was run from and pretty unrelated to that dir.