[Profiling] Multi-component programs (#2269)
This PR deals with profiling multi-component programs. It contains:

- Modifications to TDCC to produce a single json after _processing the
whole program_ (rather than outputting JSON after each component)
- A `component-cells` tool that produces information about the
user-defined component cells (so, no primitives and no constant cells)
in each component.
- Modifications to existing profiling scripts to take multi-component
programs into account. I also cleaned the scripts up a bit and left more
documentation in the comments.

The component-cells backend is necessary because the instantiation of a
component is used in the VCD, but we need to know the name of the
original component to fetch the groups that could execute in that cell.
i.e., in
[`multi-component.futil`](https://github.com/calyxir/calyx/blob/main/examples/futil/multi-component.futil),
we have `id`, which is an instantiation of the `identity()` component.
The `go` signal for the `save` group would subsequently be
`TOP.TOP.main.id.save_go...`. So, we need some way of relating that
there is a `id` cell that instantiates `identity` within the `main`
component.

Any and all feedback would be appreciated :) Additionally, if anyone has
better names for `component-cells` (or the labels in the JSON, in the
Usage section below), it would be very helpful! (Tagging @ekiwi here
since he wasn't in the list of suggested reviewers)

We are now in a nice place where we can start profiling (non-optimized)
Calyx programs! My next step is to try the profiler on some "real-life"
programs (maybe the benchmarks on the performance dashboard).

## Usage for `component-cells` tool

The tool lives in `tools/component_cells/`, so to run:
ex)
```
cargo run --manifest-path tools/component_cells/Cargo.toml examples/futil/multi-component.futil
```
will produce
```
[
  {
    "component": "main",
    "is_main_component": true,
    "cell_info": [
      {
        "cell_name": "id",
        "component_name": "identity"
      }
    ]
  },
  {
    "component": "identity",
    "is_main_component": false,
    "cell_info": []
  }
]
```
indicating that (1) `main` is the main/entry point component, (2) `main`
contains a cell named `id` that instantiates the `identity` component,
(3) the `identity` component doesn't have any (non-primitive,
non-constant) cells.