[Profiling] Differentiating FSM tracking of groups vs ground truths (go, done) (#2277)
This PR contains bug fixes and adjustments for:
1. Register-based control flow coordination taking extra cycles
2. Stream-processing VCD files are prone to bugs

To elaborate for my own sake:
1. Any register that TDCC generates to track control flow use an extra
cycle to process. For example, a group `g`'s `go` and `done` signals may
denote that `g` takes only 2 cycles. However, if an FSM register `fsm`
uses 0 to encode **when `g` is allowed to be active**, then `fsm` would
have value 0 for **3 cycles** (2 cycles for `g` + 1 cycle for reading in
that `g` is done).
To deal with this situation, the profiler maintains two versions of
information for all FSM-managed groups: (a) the ground truth information
provided by `go` and `done` signals for the group (ex.
`TOP.toplevel.main.write` in the below screenshot), and (b) the FSM
register-based information (ex. `TOP.toplevel.main.write[FSM]` in the
below screenshot). The screenshot is the summary provided for
`tests/profiler/simple-seq.futil`)

![image](https://github.com/user-attachments/assets/2ee80aea-1ec9-4e93-b67c-d80b6ab6aa72)
Having both versions will help us understand the impacts of these extra
(phantom) cycles.

2. While working on (1), I found that my VCD parsing script didn't
capture some value changes that happened simultaneously with the clock
cycle updates. The below is a small illustrative example (denoted
`timestamp signal value`):
```
70 a 1
70 clk 1
70 b 1
```
We see that on the cycle starting at timestamp 70, both `a` and `b`
updated values. However, since my script read the VCD file through a
stream, it only noticed that `a` updated values (because it was listed
before the `clk` update).
To fix this bug, I read the value changes into a dictionary from
timestamps to value change events, so that I can process all events that
happened at the same time as a cycle update.