Improve Static FSM Compilation (#2215)
Implements the ideas from #2207. 

Apologies for the gigantic PR, there are a couple reasons why it is so
big:
1) It represents a big change in the compiler 
2) I didn't know if it was going to even be worth it to implement these
changes to the compiler, so I implemented some improvements to the
compilation process that complicated the code (but
[improved](https://github.com/orgs/calyxir/discussions/2202#discussioncomment-10014608)
results).

There are some minor changes to `static_inline.rs` (in particular,
inlining `static par` blocks is more complicated now because we can't
merge always just merge all threads of a `static par` into the same
group). There are some changes to `compile_static.rs`, but the main
contribution of this PR is the file `static_tree.rs`.

I'm still going to write some tests to make sure I'm getting all edge
cases for this new tree-looking FSM compilation.