Attribute to retain cells during optimization (#2303)
I added a new boolean attribute that indicates which cells (and uses of
cells) should be preserved during optimizations. I check for the
attribute in the `dead-cell-removal` pass, so the pass doesn't optimize
those cells when it would otherwise do so. (@ekiwi and I got to the
bottom of the errors I was experiencing yesterday)

The attribute is currently called `@protected`, but if anyone has
suggestions about what the attribute should be named that would be
helpful!

## Example
The test file `tests/passes/dead-cell-removal-protected.futil` differs
from `tests/passes/dead-cell-removal.futil` by a single line, where we
add `@protected` to unused register `unused_reg`:
```
$ diff tests/passes/dead-cell-removal.futil tests/passes/dead-cell-removal-protected.futil               
30c30                                                                                                                                                    
<     unused_reg = std_reg(32);                                                                                                                          
---                                                                                                                                                      
>     @protected unused_reg = std_reg(32);                                                                                                               
```
The expect files showing the result of running `dead-cell-removal`
indicate that the pass retained `unused_reg` because of the `@protected`
attribute:
```
$ diff tests/passes/dead-cell-removal.expect tests/passes/dead-cell-removal-protected.expect             
26a27                                                                                                                                                    
>     @protected unused_reg = std_reg(32); 
```